+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 **See [the release process docs](docs/howtos/cut-a-new-release.md) for the steps to take when cutting a new release.**
 
-[Full Changelog](https://github.com/mozilla/application-services/compare/v0.22.1...master)
+[Full Changelog](https://github.com/mozilla/application-services/compare/v0.23.0...master)
+
+# v0.23.0 (_2018-03-29_)
+
+[Full Changelog](https://github.com/mozilla/application-services/compare/v0.22.1...v0.23.0)
 
 ## Places